requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
magic-filter==1.0.12
multidict==6.1.0
openai==1.52.2
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
pytz==2024.2
sniffio==1.3.1
tqdm==4.66.6
typing_extensions==4.12.2
yarl==1.17.0
